+++ deltaflow/interpolants/__init__.py
@@ -3,11 +3,13 @@
 from ..core.base_interpolant import BaseInterpolant
 from .linear import LinearInterpolant
 from .ot import OTInterpolant
+from .schrodinger_bridge import SchrodingerBridgeInterpolant
 from .variance_preserving import VariancePreservingInterpolant
 
 __all__ = [
     "BaseInterpolant",
     "LinearInterpolant",
     "OTInterpolant",
+    "SchrodingerBridgeInterpolant",
     "VariancePreservingInterpolant",
 ]